:recycle: Extract class.go
